Добавлен JSON парсер (неуниверсальный) - Изменена структура данных MediaInfo, длительность проигрывания теперь хранится в uint32_t, также были изменены типы в MP3Decoder и MediaFileHandler, и убран static_cast в GUI - Парсер добавлен в MediaFileHandler в качестве переменной класса - При неудачном считывании данных из json файла сканируется папка audio и после чего данные сохраняются в json файл - В Cmakelists добавлены новые файлы(парсер)

+++ include/media_file_info.h
@@ -7,7 +7,7 @@ struct MediaInfo
     uint64_t id;
     std::string name;
     std::string path;
-    double duration;
+    uint32_t duration;
     bool available = true;
 };
